Respect boolean parameter ylabel (disable when False)
Currently, the parameter =ylabel= is ignored and the hard-coded ylable is
written regardless of the =ylable= parameter is truthy or falsy.

This change will not set an ylable for the chart if the =ylable= parameter is
set to a falsy value.

In a next step one could check if =ylable= is a String and if so, set the actual
ylable string to that string.